Added swipe to delete alarm item

+++ app/src/main/java/com/mbexample/alarmmanager/data/repository/AlarmRepository.kt
@@ -15,6 +15,6 @@ interface AlarmRepository {
 
     suspend fun deleteAlarm(alarm: Alarm)
 
-    suspend fun deleteAlarmById(alarmId: Int)
+    suspend fun deleteAlarmById(alarmId: Long)
 
 }
